+++ src/components/popupmenu/popupmenu.js
@@ -1572,6 +1572,14 @@ PopupMenu.prototype = {
     const wrapper = this.menu.parent('.popupmenu-wrapper');
     this.wrapperPlace.setArrowPosition(e, placementObj, wrapper);
 
+    // Check if colorpicker has scrollbar
+    // Use CSS class to modify width of popupmenu.
+    if (placementObj.height < this.menu[0].offsetHeight) {
+      if (this.menu[0].classList.contains('colorpicker')) {
+        this.menu[0].classList.add('has-scrollbar');
+      }
+    }
+
     if (placementObj.height) {
       wrapper[0].style.height = '';
       this.menu[0].style.height = (placementObj.height) + (/(px|%)/i.test(`${placementObj.height}`) ? '' : 'px');